+++ src/Ema/Example/Ex02_Basic.hs
@@ -10,7 +10,7 @@ import Ema (Ema (..))
 import qualified Ema
 import qualified Ema.CLI
 import qualified Ema.CLI as CLI
-import qualified Ema.Helper.Tailwind as Tailwind
+import qualified Ema.Helper.Blaze as EB
 import Text.Blaze.Html5 ((!))
 import qualified Text.Blaze.Html5 as H
 import qualified Text.Blaze.Html5.Attributes as A
@@ -42,7 +42,7 @@ main = do
 
 render :: Some Ema.CLI.Action -> Model -> Route -> LByteString
 render emaAction model r =
-  Tailwind.layout emaAction (H.title "Basic site" >> H.base ! A.href "/") $
+  EB.twindLayout emaAction (H.title "Basic site" >> H.base ! A.href "/") $
     H.div ! A.class_ "container mx-auto" $ do
       H.div ! A.class_ "mt-8 p-2 text-center" $ do
         case r of