+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - `runEma` and friends: 
   - return the monadic's action's return value or generated files (dependent type)
 - CLI: add `run` subcommand that takes `--host` and `--port` (and remove environment hacks of $HOST and $PORT)
+- Tailwind: module revamped and renamed to `Tailwind.Helper.Blaze`
 
 ## 0.2.0.0 -- 2021-11-21